+++ bin/check-aggregate.rb
@@ -23,6 +23,12 @@ class CheckAggregate < Sensu::Plugin::Check::CLI
          description: 'Sensu API URL',
          default: 'http://localhost:4567'
 
+  option :insecure,
+         short: '-k',
+         boolean: true,
+         description: 'Enabling insecure connections',
+         default: false
+
   option :user,
          short: '-u USER',
          long: '--user USER',
@@ -66,6 +72,13 @@ class CheckAggregate < Sensu::Plugin::Check::CLI
          description: 'Summarize check result output',
          default: false
 
+  option :collect_output,
+         short: '-o',
+         long: '--output',
+         boolean: true,
+         description: 'Collects all non-ok outputs',
+         default: false
+
   option :warning,
          short: '-W PERCENT',
          long: '--warning PERCENT',
@@ -96,9 +109,12 @@ class CheckAggregate < Sensu::Plugin::Check::CLI
          description: 'A custom error MESSAGE'
 
   def api_request(resource)
+    verify_mode = OpenSSL::SSL::VERIFY_PEER
+    verify_mode = OpenSSL::SSL::VERIFY_NONE if config[:insecure]
     request = RestClient::Resource.new(config[:api] + resource, timeout: config[:timeout],
                                                                 user: config[:user],
-                                                                password: config[:password])
+                                                                password: config[:password],
+                                                                verify_ssl: verify_mode)
     JSON.parse(request.get, symbolize_names: true)
   rescue Errno::ECONNREFUSED
     warning 'Connection refused'
@@ -113,27 +129,33 @@ class CheckAggregate < Sensu::Plugin::Check::CLI
   end
 
   def honor_stash(aggregate)
-    if config[:honor_stash]
-      aggregate[:results].delete_if do |entry|
-        begin
-          api_request("/stashes/silence/#{entry[:client]}/#{config[:check]}")
-          if entry[:status] == 0
-            aggregate[:ok] = aggregate[:ok] - 1
-          elsif entry[:status] == 1
-            aggregate[:warning] = aggregate[:warning] - 1
-          elsif entry[:status] == 2
-            aggregate[:critical] = aggregate[:critical] - 1
-          else
-            aggregate[:unknown] = aggregate[:unknown] - 1
-          end
-          aggregate[:total] = aggregate[:total] - 1
-          true
-        rescue RestClient::ResourceNotFound
-          false
+    aggregate[:results].delete_if do |entry|
+      begin
+        api_request("/stashes/silence/#{entry[:client]}/#{config[:check]}")
+        if entry[:status] == 0
+          aggregate[:ok] = aggregate[:ok] - 1
+        elsif entry[:status] == 1
+          aggregate[:warning] = aggregate[:warning] - 1
+        elsif entry[:status] == 2
+          aggregate[:critical] = aggregate[:critical] - 1
+        else
+          aggregate[:unknown] = aggregate[:unknown] - 1
         end
+        aggregate[:total] = aggregate[:total] - 1
+        true
+      rescue RestClient::ResourceNotFound
+        false
       end
     end
     aggregate
+  end
+
+  def collect_output(aggregate)
+    output = ''
+    aggregate[:results].each do |entry|
+      output << entry[:output] + "\n" unless entry[:status] == 0
+    end
+    aggregate[:outputs] = [output]
   end
 
   def acquire_aggregate
@@ -145,9 +167,8 @@ class CheckAggregate < Sensu::Plugin::Check::CLI
       unless time.nil?
         uri += "/#{time}?"
         uri += '&summarize=output' if config[:summarize]
-        uri += '&results=true' if config[:honor_stash]
-        aggregate = api_request(uri)
-        honor_stash(aggregate)
+        uri += '&results=true' if config[:honor_stash] || config[:collect_output]
+        api_request(uri)
       else
         warning "No aggregates older than #{config[:age]} seconds"
       end
@@ -159,7 +180,7 @@ class CheckAggregate < Sensu::Plugin::Check::CLI
   def compare_thresholds(aggregate)
     percent_non_zero = (100 - (aggregate[:ok].to_f / aggregate[:total].to_f) * 100).to_i
     message = ''
-    if config[:summarize]
+    if aggregate[:outputs]
       aggregate[:outputs].each do |output, count|
         message << "\n" + output.to_s if count == 1
       end
@@ -200,6 +221,9 @@ class CheckAggregate < Sensu::Plugin::Check::CLI
     critical 'Misconfiguration: critical || warning || (summarize && pattern) must be set' unless threshold || pattern
 
     aggregate = acquire_aggregate
+    aggregate = honor_stash(aggregate) if config[:honor_stash]
+    puts aggregate
+    aggregate = collect_output(aggregate) if config[:collect_output]
     compare_thresholds(aggregate) if threshold
     compare_pattern(aggregate) if pattern