+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachangelog.com/)
 
 ## Unreleased
+- check-aggregates.rb adding misconfiguration check to 
+- check-aggregates.rb If summarize is set and the threshold output is being used the alert will be the summarized results
+- check-aggregates.rb Added new flag to honor stashed checks. If -i is supplied and the threshold alert is being used it will remove any checks that are stashed
 
 [0.1.0] - 2016-01-08
 ### Added